requirements.txt
forex_python
python-binance